refactor(test): 궁성 이동 방향 로직 테스트 코드 패키지 이동
